Auto increase max iterations if framerate is high enough

+++ src/fractal/lib/global_bindings.wgsl
@@ -3,7 +3,7 @@
 struct GlobalUniforms {
     time: f32,
     window_size: vec2<u32>,
-    max_iterations: u32,
+    max_iterations: f32,
     center: vec2<f32>,
     zoom: f32,
     rot: f32,
